+++ examples/main/src/App.jsx
@@ -130,7 +130,7 @@ class App extends PureComponent {
     }
     return (
       isEditorShown && (
-        <ReflexElement key={`editor-section-${this.state.viewerResetKey}`} className="section">
+        <ReflexElement key={`editor-section-${this.state.editorResetKey}`} className="section editor-example">
           <SectionHeader
             title="Editor"
             settings={settings}
@@ -169,7 +169,7 @@ class App extends PureComponent {
     const viewerState = JSON.parse(JSON.stringify(convertToRaw(editorState.getCurrentContent()))); //emulate initilState passed in by consumers
     return (
       isViewerShown && (
-        <ReflexElement key={`viewer-section-${this.state.viewerResetKey}`} className="section">
+        <ReflexElement key={`viewer-section-${this.state.viewerResetKey}`} className="section viewer-example">
           <SectionHeader
             title="Viewer"
             settings={settings}