+++ examples/main/src/viewer/ViewerPlugins.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import theme from '../theme/theme';
+import { isSSR } from 'wix-rich-content-common';
 import { videoTypeMapper } from 'wix-rich-content-plugin-video/dist/module.viewer';
 import { dividerTypeMapper } from 'wix-rich-content-plugin-divider/dist/module.viewer';
 import { HTML_TYPE, htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer';
 import { soundCloudTypeMapper } from 'wix-rich-content-plugin-sound-cloud/dist/module.viewer';
 import { LINK_TYPE, linkTypeMapper } from 'wix-rich-content-plugin-link/dist/module.viewer';
 import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer';
-import { galleryTypeMapper } from 'wix-rich-content-plugin-gallery/dist/module.viewer';
+import { galleryTypeMapper, GALLERY_TYPE } from 'wix-rich-content-plugin-gallery/dist/module.viewer';
 import { mapTypeMapper } from 'wix-rich-content-plugin-map/dist/module.viewer';
 import { giphyTypeMapper, GIPHY_TYPE } from 'wix-rich-content-plugin-giphy/dist/module.viewer';
 import { HashtagDecorator } from 'wix-rich-content-plugin-hashtag/dist/module.viewer';
@@ -73,6 +74,9 @@ export const typeMappers = [
 ];
 
 export const config = {
+  [GALLERY_TYPE]: {
+    scrollingElement: () => !isSSR() && document.getElementsByClassName('viewer-example')[0],
+  },
   [HEADERS_MARKDOWN_TYPE]: {
     hideMarkdown: true,
   },